decenter-yolov3/requirements.txt
numpy>=1.18.2
scipy>=1.4.1
tensorflow==2.3.1
tensorflow-gpu==2.3.1
hdf5storage
py3nvml
paho-mqtt
PyYAML
opencv-python
pytesseract